#6 Fix: Change UserEntity id type UserEntity의 id타입을 Integer->Long으로 모두 변경함

+++ back/tingle/src/main/java/com/example/tingle/store/dto/OrderDto.java
@@ -25,7 +25,7 @@ public class OrderDto {
     }
 
 
-    private Integer fanId;
+    private Long fanId;
     private UserDto fan;
     private ProductDto goods;
